+++ _test/test/test_integration_test.dart
@@ -15,6 +15,26 @@ void main() {
     await expectTestsPass(usePrecompiled: true);
   });
 
+  test('Can build and run a single test with --precompiled and --build-filter',
+      () async {
+    var buildArgs = [
+      '--build-filter',
+      'test/hello_world_test.*.dart.js',
+      '--build-filter',
+      'package:build_web_compilers/**'
+    ];
+    await expectTestsPass(
+        usePrecompiled: true,
+        testArgs: ['test/hello_world_test.dart'],
+        buildArgs: buildArgs);
+
+    // This wasn't built so it should fail
+    await expectTestsFail(
+        usePrecompiled: true,
+        testArgs: ['test/hello_world_custom_html_test.dart'],
+        buildArgs: buildArgs);
+  });
+
   test('Failing tests print mapped stack traces', () async {
     var result = await runTests(
         testArgs: ['--run-skipped', 'test/hello_world_test.dart']);