+++ build_runner/lib/src/entrypoint/build.dart
@@ -27,6 +27,7 @@ class BuildCommand extends BuildRunnerCommand {
     var options = readOptions();
     var result = await build(
       builderApplications,
+      buildFilters: options.buildFilters,
       deleteFilesByDefault: options.deleteFilesByDefault,
       enableLowResourcesMode: options.enableLowResourcesMode,
       configKey: options.configKey,