+++ build_runner_core/lib/src/generate/options.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'package:build/build.dart';
 import 'package:build_config/build_config.dart';
 import 'package:build_resolvers/build_resolvers.dart';
+import 'package:glob/glob.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
@@ -14,6 +15,7 @@ import 'package:path/path.dart' as p;
 import '../environment/build_environment.dart';
 import '../package_graph/package_graph.dart';
 import '../package_graph/target_graph.dart';
+import '../util/hash.dart';
 import 'exceptions.dart';
 
 /// The default list of files to include when an explicit include is not
@@ -54,21 +56,79 @@ class LogSubscription {
   final StreamSubscription<LogRecord> logListener;
 }
 
+/// Describes a set of files that should be built.
+class BuildFilter {
+  /// The package name glob that files must live under in order to match.
+  final Glob _package;
+
+  /// A glob for files under [_package] that must match.
+  final Glob _path;
+
+  BuildFilter(this._package, this._path);
+
+  /// Builds a [BuildFilter] from a command line argument.
+  ///
+  /// Both relative paths and package: uris are supported. Relative
+  /// paths are treated as relative to the [rootPackage].
+  ///
+  /// Globs are supported in package names and paths.
+  factory BuildFilter.fromArg(String arg, String rootPackage) {
+    var uri = Uri.parse(arg);
+    if (uri.scheme == 'package') {
+      var package = uri.pathSegments.first;
+      var glob = Glob(p.url.joinAll([
+        'lib',
+        ...uri.pathSegments.skip(1),
+      ]));
+      return BuildFilter(Glob(package), glob);
+    } else if (uri.scheme.isEmpty) {
+      return BuildFilter(Glob(rootPackage), Glob(uri.path));
+    } else {
+      throw FormatException('Unsupported scheme ${uri.scheme}', uri);
+    }
+  }
+
+  /// Returns whether or not [id] mathes this filter.
+  bool matches(AssetId id) =>
+      _package.matches(id.package) && _path.matches(id.path);
+
+  @override
+  int get hashCode {
+    var hash = 0;
+    hash = hashCombine(hash, _package.context.hashCode);
+    hash = hashCombine(hash, _package.pattern.hashCode);
+    hash = hashCombine(hash, _package.recursive.hashCode);
+    hash = hashCombine(hash, _path.context.hashCode);
+    hash = hashCombine(hash, _path.pattern.hashCode);
+    hash = hashCombine(hash, _path.recursive.hashCode);
+    return hashComplete(hash);
+  }
+
+  @override
+  operator ==(other) =>
+      other is BuildFilter &&
+      other._path.context == _path.context &&
+      other._path.pattern == _path.pattern &&
+      other._path.recursive == _path.recursive &&
+      other._package.context == _package.context &&
+      other._package.pattern == _package.pattern &&
+      other._package.recursive == _package.recursive;
+}
+
 /// Manages setting up consistent defaults for all options and build modes.
 class BuildOptions {
-  // Build mode options.
-  final StreamSubscription logListener;
-  final PackageGraph packageGraph;
-
   final bool deleteFilesByDefault;
   final bool enableLowResourcesMode;
-  final bool trackPerformance;
-  final bool verbose;
-  final TargetGraph targetGraph;
-  final Resolvers resolvers;
+  final StreamSubscription logListener;
 
   /// If present, the path to a directory to write performance logs to.
   final String logPerformanceDir;
+
+  final PackageGraph packageGraph;
+  final Resolvers resolvers;
+  final TargetGraph targetGraph;
+  final bool trackPerformance;
+  final bool verbose;
 
   // Watch mode options.
   Duration debounceDelay;